+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 ## 1.1.0 (TBD)
 
-* Add CallableMetadata.returnType
+* Support property metadata with ClassMetadata.getProperty
+* Allow to enumerate all declarations in a class with ClassMetadata.{functions,constructors,properties}
+* Add CallableMetadata.{name,returnType,extensionReceiverType}
+* Add FunctionMetadata.{isInline,isExternal,isOperator,isInfix,isSuspend}
+* Add PropertyMetadata.{isLateinit,isConst}
+* Add ClassMetadata.isData
+* Add ClassMetadata.kind
+* Add ConstructorMetadata.isPrimary
+* Add DeclarationMetadata.visibility
+* Add ParameterMetadata.hasDefaultValue
+* Update to Kotlin 1.2.0
 
 ## 1.0.0 (Feb 15, 2016)